+++ components/Earth.tsx
@@ -2,7 +2,7 @@ import React, { Suspense } from "react";
 import { Canvas } from "@react-three/fiber";
 import { OrbitControls, Preload, useGLTF } from "@react-three/drei";
 
-import {CanvasLoader} from "../components";
+import {CanvasLoader} from ".";
 
 const Earth = () => {
   const earth = useGLTF("./planet/scene.gltf");
