+++ components/StartSteps.tsx
@@ -1,6 +1,6 @@
 import styles from '../styles';
 
-const StartSteps = ({ number, text }) => (
+const StartSteps = ({ number, text } : any) => (
   <div className={`${styles.flexCenter} flex-row`}>
     <div
       className={`${styles.flexCenter} w-[70px] h-[70px] rounded-[24px] bg-[#323F5D]`}
