--- sections/About.jsx
@@ -1,44 +1,0 @@
-'use client';
-
-import { motion } from 'framer-motion';
-import { TypingText } from '../components';
-
-import styles from '../styles';
-import { fadeIn, staggerContainer } from '../utils/motion';
-
-const About = () => (
-  <section className={`${styles.paddings} relative z-10`}>
-    <div className="gradient-02 z-0" />
-    <motion.div
-      variants={staggerContainer}
-      initial="hidden"
-      whileInView="show"
-      viewport={{ once: true, amount: 0.25 }}
-      className={`${styles.innerWidth} mx-auto ${styles.flexCenter} flex-col`}
-    >
-      <TypingText title="| About" textStyles="text-center" />
-
-      <motion.p
-        variants={fadeIn('up', 'tween', 0.2, 1)}
-        className="mt-[8px] font-normal sm:text-[32px] text-[20px] text-center text-secondary-white"
-      >
-        <span className="font-extrabold text-white">CDS</span>  an open-source community on a mission to address real-world challenges through our cutting-edge projects. Committed to open-source values, we champion collaboration and remote work, enabling coders from diverse backgrounds to unite and make a significant difference.Let’s explore the {' '}
-        <span className="font-extrabold text-white">
-          madness of the CDS
-        </span>{' '}
-        
-        
-        by scrolling down !
-      </motion.p>
-
-      <motion.img
-        variants={fadeIn('up', 'tween', 0.3, 1)}
-        src="/arrow-down.svg"
-        alt="arrow down"
-        className="w-[18px] h-[28px] object-contain mt-[28px]"
-      />
-    </motion.div>
-  </section>
-);
-
-export default About;
