+++ sections/CodeOfConduct.tsx
