+++ sections/ContiGuide.tsx
