+++ sections/FounderNote.tsx
