+++ sections/Hero.tsx
