+++ tsconfig.json
@@ -17,6 +17,6 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "components/CustomTexts.jsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "components/CustomTexts.tsx"],
   "exclude": ["node_modules"]
 }
